requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.13.0
PyJWT==2.8.0
requests==2.31.0
gunicorn==21.2.0